gps data change
