+++ safe/app/build.gradle
@@ -60,7 +60,7 @@ dependencies {
     //server
     implementation 'com.squareup.okhttp3:okhttp:4.9.1'
 
-    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.appcompat:appcompat:1.6.0'
     implementation 'com.google.android.material:material:1.8.0'
     implementation 'com.google.android.gms:play-services-maps:18.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
